backend/requirements.txt
django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.2
scrapy==2.11.1
python-dotenv==1.0.1
requests==2.31.0